Add preptime
I want the user to know how long it will
take to process their order.

Each item in the store has a preparation time,
defaulting to 12 minutes. Items can be edited to take longer.
If an order has more than six items, add 10 minutes
for every additional six items.
Each already "paid" order in the system which is not "complete"
delays the production start of this new order by 4 minutes.